fixed bug with comment eventlistener, does not add click action when movie list is presented from search input
