+++ src/JS_Modules/Controllers/PresentUI.js
@@ -2,7 +2,6 @@ import commentsModal from '../Views/Comments.js';
 import footer from '../Views/Footer.js';
 import header from '../Views/Header.js';
 import moviesList from '../Views/MoviesList.js';
-import commentsModalHandler from './ModalHandler.js';
 import MoviesListManager from './MovieListManager.js';
 import searchBar from './SearchBar.js';
 
@@ -14,14 +13,6 @@ const presentUI = async () => {
   document.body.insertAdjacentHTML('beforeend', commentsModal());
 
   await MoviesListManager();
-
-  const commentsModalBtn = document.querySelectorAll('.commentsModalBtn');
-  commentsModalBtn.forEach((button) => {
-    const moveID = button.getAttribute('data-movie-id');
-    button.addEventListener('click', () => {
-      commentsModalHandler(moveID);
-    });
-  });
 
   const searchForm = document.querySelector('.searchMovie');
   searchForm.addEventListener('submit', (e) => { e.preventDefault(); searchBar(searchForm); });
